requirements.txt
# Data
pandas>=2.2.0
requests>=2.32.0

# Database
psycopg2-binary>=2.9.9

# GCP
google-cloud-storage>=3.0.0

# Airflow
apache-airflow>=3.0.0

# DVC
dvc>=3.0.0
dvc-gs>=3.0.0

# Web Scraping
beautifulsoup4>=4.12.0

# LangChain
langchain-core>=0.1.0
langchain-huggingface>=0.0.1
langchain-pinecone>=0.0.1

# Pinecone
pinecone-client>=3.0.0

# PDF Processing
pymupdf>=1.23.0
langchain-experimental>=0.0.1

# Bias Detection
fairlearn>=0.9.0

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0